✨ feat(cast_of_structures): Add casting between structs in C
Allows interpreting the same block of memory in different ways using structs. It also introduces the concept of endianness, which refers to the order in which bytes are stored in memory. The code includes examples of casting between two structs with different sizes and checking the endianness of the system.